+++ truongdev/src/main/java/com/dev/truongdev/api/ProjectAPI.java
@@ -1,7 +1,7 @@
 package com.dev.truongdev.api;
 
-import com.dev.truongdev.dto.dashload.department.DepartmentStatsDTO;
-import com.dev.truongdev.dto.dashload.project.ProjectStatsDTO;
+import com.dev.truongdev.dto.dashboard.department.DepartmentStatsDTO;
+import com.dev.truongdev.dto.dashboard.project.ProjectStatsDTO;
 import com.dev.truongdev.entity.Project;
 import com.dev.truongdev.entity.ProjectHistory;
 import com.dev.truongdev.dto.ProjectDTO;
